Move RepresentationResolver to separate file
