+++ pennylane/circuit_drawer/circuit_drawer.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Xanadu Quantum Technologies Inc.
+# Copyright 2019-2020 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ import numpy as np
 import pennylane as qml
 
 from .charsets import UnicodeCharSet
+from .representation_resolver import RepresentationResolver
 
 # pylint: disable=too-many-branches,too-many-arguments,too-many-return-statements,too-many-statements,consider-using-enumerate,too-many-instance-attributes
 
@@ -185,339 +186,6 @@ class Grid:
             ret += "\n"
 
         return ret
-
-class RepresentationResolver:
-    """Resolves the string representation of PennyLane objects.
-
-    Args:
-        charset (CharSet, optional): The CharSet to be used for representation resolution.
-        show_variable_names (bool, optional): Show variable names instead of variable values.
-    """
-
-    # Symbol for uncontrolled wires
-    resolution_dict = {
-        "PauliX": "X",
-        "CNOT": "X",
-        "Toffoli": "X",
-        "CSWAP": "SWAP",
-        "PauliY": "Y",
-        "PauliZ": "Z",
-        "CZ": "Z",
-        "Identity": "I",
-        "Hadamard": "H",
-        "CRX": "RX",
-        "CRY": "RY",
-        "CRZ": "RZ",
-        "CRot": "Rot",
-        "PhaseShift": "Rϕ",
-        "Beamsplitter": "BS",
-        "Squeezing": "S",
-        "TwoModeSqueezing": "S",
-        "Displacement": "D",
-        "NumberOperator": "n",
-        "Rotation": "R",
-        "ControlledAddition": "X",
-        "ControlledPhase": "Z",
-        "ThermalState": "Thermal",
-        "GaussianState": "Gaussian",
-        "QuadraticPhase": "P",
-        "CubicPhase": "V",
-        "X": "x",
-        "P": "p",
-    }
-    """Symbol used for uncontrolled wires."""
-
-    control_wire_dict = {
-        "CNOT": [0],
-        "Toffoli": [0, 1],
-        "CSWAP": [0],
-        "CRX": [0],
-        "CRY": [0],
-        "CRZ": [0],
-        "CRot": [0],
-        "CZ": [0],
-        "ControlledAddition": [0],
-        "ControlledPhase": [0],
-    }
-    """Indices of control wires."""
-
-    def __init__(self, charset=UnicodeCharSet, show_variable_names=False):
-        self.charset = charset
-        self.show_variable_names = show_variable_names
-        self.matrix_cache = []
-        self.unitary_matrix_cache = []
-        self.hermitian_matrix_cache = []
-
-    @staticmethod
-    def index_of_array_or_append(target_element, target_list):
-        """Returns the first index of an appearance of the target element in the target list.
-        If the target element is not in the list it will be added to the list.
-
-        Args:
-            target_element (np.ndarray): The object whose index is to be returned
-            target_list (list[np.ndarray]): The list which shall be searched
-
-        Returns:
-            int: Index of the target element in the list.
-        """
-        for idx, target in enumerate(target_list):
-            if np.array_equal(target, target_element):
-                return idx
-
-        target_list.append(target_element)
-
-        return len(target_list) - 1
-
-    def single_parameter_representation(self, par):
-        """Resolve the representation of an Operator's parameter.
-
-        Args:
-            par (Union[qml.variable.Variable, int, float]): The parameter to be rendered
-
-        Returns:
-            str: String representation of the parameter
-        """
-        if isinstance(par, qml.variable.Variable):
-            return par.render(self.show_variable_names)
-
-        return str(round(par, 3))
-
-    @staticmethod
-    def _format_poly_term(coefficient, variable):
-        """Format a term of a polynomial."""
-        if coefficient == 0:
-            return ""
-
-        if coefficient == 1.0:
-            return str(variable)
-
-        if coefficient == -1.0:
-            return "-" + str(variable)
-
-        return "{:+.3g}{}".format(
-            coefficient, variable
-        )
-
-    def operator_representation(self, op, wire):
-        """Resolve the representation of an Operator.
-
-        Args:
-            op (pennylane.operation.Operator): The Operator instance whose representation shall be resolved
-            wire (int): The Operator's wire which is currently resolved
-
-        Returns:
-            str: String representation of the Operator
-        """
-        if isinstance(op, qml.operation.Tensor):
-            constituent_representations = [
-                self.operator_representation(tensor_obs, wire) for tensor_obs in op.obs
-            ]
-
-            return (" " + self.charset.OTIMES + " ").join(constituent_representations)
-
-        name = op.name
-
-        # Use a shorter name if applicable
-        if name in RepresentationResolver.resolution_dict:
-            name = RepresentationResolver.resolution_dict[name]
-
-        # Display a control symbol for all controlling qubits of a controlled Operation
-        if op.name in self.control_wire_dict and wire in [
-            op.wires[control_idx] for control_idx in self.control_wire_dict[op.name]
-        ]:
-            return self.charset.CONTROL
-
-        if op.num_params == 0:
-            return name
-
-        if op.name == "QubitUnitary":
-            mat = op.params[0]
-            idx = RepresentationResolver.index_of_array_or_append(mat, self.unitary_matrix_cache)
-
-            return "U{}".format(idx)
-
-        if op.name == "Hermitian":
-            mat = op.params[0]
-            idx = RepresentationResolver.index_of_array_or_append(mat, self.hermitian_matrix_cache)
-
-            return "H{}".format(idx)
-
-        if op.name == "QuadOperator":
-            par_rep = self.single_parameter_representation(op.params[0])
-
-            return "cos({0})x+sin({0})p".format(par_rep)
-
-        if op.name == "FockStateProjector":
-            n_str = ",".join([str(n) for n in op.params[0]])
-
-            return (
-                self.charset.PIPE + n_str + self.charset.CROSSED_LINES + n_str + self.charset.PIPE
-            )
-
-        if op.name == "PolyXP":
-            coefficients = op.params[0]
-            order = len(coefficients.shape)
-
-            if order == 1:
-                poly_str = ""
-                if coefficients[0] != 0:
-                    poly_str += "{:.3g}".format(coefficients[0])
-
-                for idx in range(0, coefficients.shape[0] // 2):
-                    x = 2 * idx + 1
-                    y = 2 * idx + 2
-                    poly_str += RepresentationResolver._format_poly_term(
-                        coefficients[x], "x{}".format(self.charset.to_subscript(idx))
-                    )
-                    poly_str += RepresentationResolver._format_poly_term(
-                        coefficients[y], "p{}".format(self.charset.to_subscript(idx))
-                    )
-
-                return poly_str
-
-            if order == 2:
-                poly_str = str(coefficients[0, 0])
-
-                for idx in range(0, coefficients.shape[0] // 2):
-                    x = 2 * idx + 1
-                    p = 2 * idx + 2
-                    poly_str += RepresentationResolver._format_poly_term(
-                        coefficients[0, x] + coefficients[x, 0],
-                        "x{}".format(self.charset.to_subscript(idx)),
-                    )
-                    poly_str += RepresentationResolver._format_poly_term(
-                        coefficients[0, p] + coefficients[p, 0],
-                        "p{}".format(self.charset.to_subscript(idx)),
-                    )
-
-                for idx1 in range(0, coefficients.shape[0] // 2):
-                    for idx2 in range(idx1, coefficients.shape[0] // 2):
-                        x1 = 2 * idx1 + 1
-                        p1 = 2 * idx1 + 2
-                        x2 = 2 * idx2 + 1
-                        p2 = 2 * idx2 + 2
-
-                        if idx1 == idx2:
-                            poly_str += RepresentationResolver._format_poly_term(
-                                coefficients[x1, x1],
-                                "x{}{}".format(
-                                    self.charset.to_subscript(idx1), self.charset.to_superscript(2)
-                                ),
-                            )
-                            poly_str += RepresentationResolver._format_poly_term(
-                                coefficients[p1, p1],
-                                "p{}{}".format(
-                                    self.charset.to_subscript(idx1), self.charset.to_superscript(2)
-                                ),
-                            )
-                            poly_str += RepresentationResolver._format_poly_term(
-                                coefficients[x1, p1] + coefficients[p1, x1],
-                                "x{}p{}".format(
-                                    self.charset.to_subscript(idx1), self.charset.to_subscript(idx1)
-                                ),
-                            )
-                        else:
-                            poly_str += RepresentationResolver._format_poly_term(
-                                coefficients[x1, x2] + coefficients[x2, x1],
-                                "x{}x{}".format(
-                                    self.charset.to_subscript(idx1), self.charset.to_subscript(idx2)
-                                ),
-                            )
-
-                            poly_str += RepresentationResolver._format_poly_term(
-                                coefficients[p1, p2] + coefficients[p2, p1],
-                                "p{}p{}".format(
-                                    self.charset.to_subscript(idx1), self.charset.to_subscript(idx2)
-                                ),
-                            )
-
-                            poly_str += RepresentationResolver._format_poly_term(
-                                coefficients[x1, p2] + coefficients[p2, x1],
-                                "x{}p{}".format(
-                                    self.charset.to_subscript(idx1), self.charset.to_subscript(idx2)
-                                ),
-                            )
-
-                            poly_str += RepresentationResolver._format_poly_term(
-                                coefficients[p1, x2] + coefficients[x2, p1],
-                                "x{}p{}".format(
-                                    self.charset.to_subscript(idx2), self.charset.to_subscript(idx1)
-                                ),
-                            )
-
-                return poly_str
-
-        if op.name == "FockState":
-            return self.charset.PIPE + str(op.params[0]) + self.charset.RANGLE
-
-        if op.name in {"BasisState", "FockStateVector"}:
-            return self.charset.PIPE + str(op.params[0][op.wires.index(wire)]) + self.charset.RANGLE
-
-        # Operations that only have matrix arguments
-        if op.name in {
-            "GaussianState",
-            "FockDensityMatrix",
-            "FockStateVector",
-            "QubitStateVector",
-            "Interferometer",
-        }:
-            param_strings = []
-            for param in op.params:
-                idx = RepresentationResolver.index_of_array_or_append(param, self.matrix_cache)
-
-                param_strings.append("M{}".format(idx))
-
-            return "{}({})".format(name, ", ".join(param_strings))
-
-        return "{}({})".format(
-            name, ", ".join([self.single_parameter_representation(par) for par in op.params])
-        )
-
-    def output_representation(self, obs, wire):
-        """Resolve the representation of a circuit's output.
-
-        Args:
-            obs (pennylane.ops.Observable): The Observable instance whose representation shall be resolved
-            wire (int): The Observable's wire which is currently resolved
-
-        Returns:
-            str: String representation of the Observable
-        """
-        # pylint: disable=inconsistent-return-statements
-        if obs.return_type == qml.operation.Expectation:
-            return (
-                self.charset.LANGLE
-                + "{}".format(self.operator_representation(obs, wire))
-                + self.charset.RANGLE
-            )
-
-        if obs.return_type == qml.operation.Variance:
-            return "Var[{}]".format(self.operator_representation(obs, wire))
-
-        if obs.return_type == qml.operation.Sample:
-            return "Sample[{}]".format(self.operator_representation(obs, wire))
-
-        # Unknown return_type
-        return "{}[{}]".format(str(obs.return_type), self.operator_representation(obs, wire))
-
-    def element_representation(self, element, wire):
-        """Resolve the representation of an element in the circuit's Grid.
-
-        Args:
-            element (Union[NoneType,str,qml.operation.Operator]): The circuit element whose representation shall be resolved
-            wire (int): The element's wire which is currently resolved
-
-        Returns:
-            str: String representation of the element
-        """
-        if element is None:
-            return ""
-        if isinstance(element, str):
-            return element
-        if isinstance(element, qml.operation.Observable) and element.return_type is not None:
-            return self.output_representation(element, wire)
-
-        return self.operator_representation(element, wire)
 
 
 def _remove_duplicates(input_list):
